+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v12.9
 --------------------------
+- Added formula advisor
 - Added windows and macos executors
 - Localise date across the app
 - Removed time from created and updated fields